refactor: move project ownership authorization to authz service
