+++ apps/backend/src/services/authz.service.ts
@@ -362,6 +362,37 @@ export async function hasProjectRole(
 }
 
 /**
+ * Require that a user is the owner of a project
+ * Throws NotFoundError if the project doesn't exist
+ * Throws ForbiddenError if the user is not the owner
+ *
+ * @param projectId - The project ID to check ownership for
+ * @param userId - The user ID to check
+ * @throws NotFoundError if project doesn't exist
+ * @throws ForbiddenError if user is not the owner
+ */
+export async function requireProjectOwnership(
+  projectId: string,
+  userId: string
+): Promise<void> {
+  const db = getDb();
+
+  const [project] = await db
+    .select({ userId: projects.userId })
+    .from(projects)
+    .where(eq(projects.id, projectId))
+    .limit(1);
+
+  if (!project) {
+    throw new NotFoundError("Project");
+  }
+
+  if (project.userId !== userId) {
+    throw new ForbiddenError("You do not have access to this project");
+  }
+}
+
+/**
  * Require that a user has a specific role or higher for a project
  * Throws ForbiddenError if the user lacks the required role
  *
